Refactor remove container into dockerUtilities
